5,6,7接口

+++ boxfish-online-card/src/main/java/com/boxfishedu/workorder/service/StStudentApplyRecordsService.java
@@ -36,7 +36,7 @@ public class StStudentApplyRecordsService extends BaseService<StStudentApplyReco
 
 
     public Integer getUnreadInvitedNum(Long teacherId, Date date){
-       Optional<Integer> unread = jpa.getUnreadInvitedNum(teacherId,date, StStudentApplyRecords.ReadStatus.no);
+       Optional<Long> unread = jpa.getUnreadInvitedNum(teacherId,date, StStudentApplyRecords.ReadStatus.no);
         return unread.isPresent()?unread.get().intValue():0;
     }
